HOTT-3177: Refactor shared report behaviour
