+++ app/lib/reporting/declarable_duties.rb
@@ -1,5 +1,9 @@
+require 'reporting/concerns/reportable'
+
 module Reporting
   class DeclarableDuties
+    extend Reporting::Reportable
+
     MEASURE_EAGER = [
       :base_regulation,
       :modification_regulation,
@@ -206,34 +210,8 @@ module Reporting
         end
       end
 
-      def object
-        bucket.object(object_key)
-      end
-
-      def object_key
-        "#{service}/reporting/#{year}/#{month}/#{day}/declarable_commodities_with_duty_measures_#{service}_#{now.strftime('%Y_%m_%d')}.xlsx"
-      end
-
-      def bucket
-        Rails.application.config.reporting_bucket
-      end
-
-      delegate :service, to: TradeTariffBackend
-
-      def day
-        now.day.to_s.rjust(2, '0')
-      end
-
-      def month
-        now.month.to_s.rjust(2, '0')
-      end
-
-      def year
-        now.year
-      end
-
-      def now
-        Time.zone.today
+      def report_name
+        'declarable_commodities_with_duty_measures'
       end
     end
   end
